sf2_software_bin/requirements.txt
PyQt6
openpyxl
textual
textual-dev
simple-term-menu
Pillow
Flask
python-dotenv
